requirements.txt
Django==1.5.1
Jinja2==2.6
PIL==1.1.7
Whoosh==2.4.1
distribute==0.6.34
django-haystack==2.0.0-beta
gunicorn==0.17.2
heroku==0.1.2
honcho==0.4.0
oauthlib==0.4.0
pycrypto==2.6
python-dateutil==1.5
requests==1.2.0
requests-oauthlib==0.3.0
smartfile==2.1-1
virtualenv==1.9.1
